// webapp/src/app/Model/Resource.ts
export class Resource {
    
    
    id:Number;
    type: String;
    bedType: String;
    medicineName:String;
    typeofEquipment:String;
    medicalEquipmentName:String;
    avalabilityPlace: String;
    address: String;
    contactPersonName: String;
    contactMobileNumber: String;
    isVerified: boolean=false;
    verifiedBy: String;
}